+++ HSbotRunner.sh
@@ -1,3 +1,3 @@
-python3 -m venv MFB
+python3.11 -m venv MFB
 ./MFB/bin/python prereqs_linux.py
 ./MFB/bin/python main.py